fixed splitting
